Pull out API tasks into a module
